added migration due to balance type change

+++ server/UserService/UserService.Data/Entities/Account.cs
@@ -9,7 +9,7 @@ namespace UserService.Data.Entities
         public Guid Id { get; set; }
         public Guid UserId { get; set; }
         public DateTime OpenDate { get; set; }
-        public float Balance { get; set; }
+        public int Balance { get; set; }
         public virtual User user { get; set; }
 
     }
