+++ server/UserService/UserService.Services/Models/AccountModel.cs
@@ -9,6 +9,6 @@ namespace UserService.Services.Models
         public Guid Id { get; set; }
         public Guid UserId { get; set; }
         public DateTime OpenDate { get; set; }
-        public float Balance { get; set; }
+        public int Balance { get; set; }
     }
 }
